pyproject.toml
[tool.poetry]
name = "postgres_to_mongo_importer"
version = "0.1.0"
description = ""
authors = ["tornike <tornike.skhulukhia.1@iliauni.edu.ge>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.2.0"
psycopg2 = "^2.9.3"
rich = "^12.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
